refactor(dto): split portfolio contracts into request and response
