+++ core/src/main/java/com/marmitt/core/dto/portfolio/request/ReprocessDeadLetterRequest.java
@@ -1,7 +1,8 @@
-package com.marmitt.core.dto.portfolio;
+package com.marmitt.core.dto.portfolio.request;
 
 public record ReprocessDeadLetterRequest(
         String requestedBy,
         String resolutionNote
 ) {
 }
+
