+++ core/src/main/java/com/marmitt/core/dto/portfolio/request/ResolveDeadLetterRequest.java
@@ -1,7 +1,8 @@
-package com.marmitt.core.dto.portfolio;
+package com.marmitt.core.dto.portfolio.request;
 
 public record ResolveDeadLetterRequest(
         String resolvedBy,
         String resolutionNote
 ) {
 }
+
