+++ core/src/main/java/com/marmitt/core/dto/portfolio/response/PortfolioPnlDto.java
@@ -1,4 +1,4 @@
-package com.marmitt.core.dto.portfolio;
+package com.marmitt.core.dto.portfolio.response;
 
 import lombok.Builder;
 
@@ -13,3 +13,4 @@ public record PortfolioPnlDto(
         BigDecimal totalPnL
 ) {
 }
+
