+++ core/src/main/java/com/marmitt/core/ports/inbound/runner/QueryRunnerPort.java
@@ -1,6 +1,6 @@
 package com.marmitt.core.ports.inbound.runner;
 
-import com.marmitt.core.dto.portfolio.TransactionDto;
+import com.marmitt.core.dto.portfolio.response.TransactionDto;
 import com.marmitt.core.dto.runner.RunnerDto;
 
 import java.util.List;
@@ -12,3 +12,4 @@ public interface QueryRunnerPort {
 
     List<TransactionDto> findTransactionsByRunnerId(UUID runnerId);
 }
+
